Code cleanup

+++ Src/Hardware/Matrix8x8.cpp
@@ -11,7 +11,7 @@ Matrix8x8::Matrix8x8(uint16_t _address) :
 
 void Matrix8x8::setBitmap(const std::array<uint8_t, 8> *_image)
 {
-	for (uint8_t i = 0; i < _image->size(); ++i) displayBuffer[i] = std::rotr((*_image)[i], 1);
+	for (unsigned int i = 0; i < _image->size(); ++i) displayBuffer[i] = std::rotr((*_image)[i], 1);
 }
 
 void Matrix8x8::setPixel(const uint8_t _x, const uint8_t _y, const bool _value)
@@ -38,7 +38,7 @@ void Matrix8x8::scroll(Direction _direction, uint8_t _value)
 		case Direction::DOWN :
 		{
 			const uint8_t shift = _direction == Direction::UP ? _value % 8 : 8 - _value % 8;
-			const int8_t gcd = std::gcd(shift, 8);
+			const uint8_t gcd = std::gcd(shift, 8);
 
 			for (uint8_t i = 0; i < gcd; i++)
 			{
